+++ conf/application.conf
@@ -33,7 +33,7 @@ play.modules.enabled += "uk.gov.hmrc.play.bootstrap.graphite.GraphiteMetricsModu
 
 # Provides an implementation and configures all filters required by a Platform frontend microservice.
 play.modules.enabled += "uk.gov.hmrc.play.bootstrap.FrontendModule"
-play.http.filters = "uk.gov.hmrc.play.bootstrap.filters.FrontendFilters"
+play.http.filters = "config.filters.ServiceFilters"
 
 # Default http client
 play.modules.enabled += "uk.gov.hmrc.play.bootstrap.AuditModule"
