+++ project/AppDependencies.scala
@@ -6,7 +6,7 @@ import sbt._
 object AppDependencies {
 
   val compile = Seq(
-    "uk.gov.hmrc"             %% "play-whitelist-filter"    % "1.1.0",
+    "uk.gov.hmrc"             %% "play-whitelist-filter"    % "2.0.0",
     "uk.gov.hmrc"             %% "govuk-template"           % "5.26.0-play-25",
     "uk.gov.hmrc"             %% "play-ui"                  % "7.27.0-play-25",
     "uk.gov.hmrc"             %% "bootstrap-play-25"        % "4.6.0"
